pyproject.toml
[tool.poetry]
name = "ecohicalrt"
version = "0.2.0"
description = "Reproducibility package for EcoHiCaLRT experiments (bug-fix + Optuna isolation)."
authors = ["AI Researcher <ai@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.2.0"
transformers = "^4.35.0"
datasets = "^2.16.0"
peft = "^0.7.0"
bitsandbytes = "^0.43.0"          # optional – LoRA 8-bit support
hydra-core = "^1.3.2"
wandb = "^0.16.0"
optuna = "^3.5.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
pandas = "^2.1.0"
scikit-learn = "^1.4.0"
scipy = "^1.11.0"
tqdm = "^4.66.0"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
